requirements.txt
mteb
python-dotenv
opensearch-py
datasets
beir